+++ .gitignore
@@ -1,3 +1,4 @@
-.idea
-vendor
+.idea/
+vendor/
 composer.phar
+composer.lock
